refactor(jb): audit fixes - deadband param, episode naming, tail ring type
Code review findings, all behavior-preserving (full suite green):
- Remove TargetControllerParams::deadband_slots (ADR-4 structurally
  enforced: nonzero deadband +跌侧grind = permanent offset; the param was
  a footgun nobody sets). Rise is now plain desired > current_; the dead
  else-branch goes away; TargetPath::Deadband enum value stays for the
  JNI/Kotlin contract (documented unreachable).
- Rename EpisodeDir::Up/Down to Filling/Dropping (1:1 with public
  EpisodeState); publish_episode_state becomes a static_cast pinned by
  static_assert instead of a ternary.
- Rename last_underrun_events_ to last_penalty_events_ (it stores the
  select_penalty_events-mapped count, not raw underruns).
- tail_ring_ double[2048] -> uint32_t[2048] (bucket indices were doubles;
  16KB -> 8KB, drops two casts per packet).

+++ aqua_core/src/audio/buffer/jitter_estimator.cpp
@@ -293,12 +293,12 @@ void JitterEstimator::observe(std::uint16_t seq, std::uint32_t timestamp, std::u
             ++tail_count_;
         } else {
             // 满窗：先退役最老样本，窗口滑出即遗忘（跌慢的来源）。
-            const auto old = static_cast<std::uint32_t>(tail_ring_[tail_head_]);
+            const auto old = tail_ring_[tail_head_];
             if (old <= config::JB_TAIL_HISTOGRAM_BUCKETS) {
                 --tail_hist_[old];
             }
         }
-        tail_ring_[tail_head_] = static_cast<double>(bucket);
+        tail_ring_[tail_head_] = bucket;
         ++tail_hist_[bucket];
         tail_head_ = (tail_head_ + 1) % config::JB_TAIL_WINDOW_PACKETS;
         // 分位点现算：64 桶线性扫找 ceil(q×n) 累积点（每包 O(64)，可忽略）。
